% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event.R
\name{new_event}
\alias{new_event}
\alias{validate_event}
\alias{event}
\title{Event Constructor}
\usage{
new_event(id, type, time, ...)

validate_event(id, type, time, ...)

event(id, type, time, ...)
}
\arguments{
\item{id}{A character identifier}

\item{type}{A character indicating the event type}

\item{time}{A character strign indicating the event timestamp}

\item{...}{A list with the event attributes.}
}
\description{
Event constructor
}
\examples{

birth_event <- event(
  id = 'first-id',
  type = 'BIRTH',
  time = '1936-11-09',
  birth_date = '1936-11-09'
)

death_event <- event(
  id = 'second-id',
  type = 'DEATH',
  time = '2019-05-22',
  death_date = '2019-05-22'
)

}
