% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_body.R
\name{get_body}
\alias{get_body}
\title{Get Event Body}
\usage{
get_body(obj)
}
\arguments{
\item{obj}{An event or a list of events.}
}
\value{
The \code{get_body} function returns a list object
}
\description{
This function returns the body of an event or list of events.
}
\examples{

first_event <- event(
  id = 'first-event',
  type = 'FIRST_EVENT',
  time = Sys.time(),
  attr_01 = 'first-body-attr',
  attr_02 = 'second-body-attr'
)

get_body(first_event)

}
