% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_header.R
\name{get_header}
\alias{get_header}
\title{Get Event Header}
\usage{
get_header(obj)
}
\arguments{
\item{obj}{An event or a list of events.}
}
\description{
This function returns the header of an event or list of events.
}
\examples{

first_event <- event(id = 'first-event', type = 'FIRST_EVENT', time = Sys.time())

get_header(first_event)

}
