% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_id.R
\name{get_id}
\alias{get_id}
\title{Get Event Identifier}
\usage{
get_id(obj)
}
\arguments{
\item{obj}{An event or a list of events.}
}
\value{
\code{get_id} returns a character string.
}
\description{
This function returns the identifier of an event or list of events.
}
\examples{

first_event <- event(id = 'first-event', type = 'FIRST_EVENT', time = Sys.time())

get_id(first_event)

}
