% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_time.R
\name{get_time}
\alias{get_time}
\title{Get Event Time}
\usage{
get_time(obj)
}
\arguments{
\item{obj}{An event or a list of events.}
}
\value{
\code{get_time()} return a POSIXct object
}
\description{
This function returns the time of an event or list of events.
}
\examples{

first_event <- event(
  id = 'first-event',
  type = 'FIRST_EVENT',
  time = Sys.time()
)

get_time(first_event)

}
