% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_type.R
\name{get_type}
\alias{get_type}
\title{Get Event Type}
\usage{
get_type(obj)
}
\arguments{
\item{obj}{An event or a list of events.}
}
\value{
\code{get_type()} return a character string
}
\description{
This function returns the type of an event or list of events.
}
\examples{

first_event <- event(id = 'first-event', type = 'FIRST_EVENT', time = Sys.time())
second_event <- event(id = 'second-event', type = 'SECOND_EVENT', time = Sys.time())

the_event_list <- event_list(first_event, second_event)

get_type(the_event_list)

}
