% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_.R
\name{is_event}
\alias{is_event}
\alias{is_event_list}
\alias{is_handlers_list}
\alias{is_handler}
\alias{is_dispatcher}
\title{Is an Object from a class?}
\usage{
is_event(x)

is_event_list(x)

is_handlers_list(x)

is_handler(x)

is_dispatcher(x)
}
\arguments{
\item{x}{any `R` object.}
}
\value{
Returns a logical value
}
\description{
Test if an object is some of the classes implemented in `eventr` package.
}
\examples{

birth_event <- event(
  id = 'first-id',
  type = 'BIRTH',
  time = '1936-11-09',
  birth_date = '1936-11-09'
)

is_event(birth_event)

}
